\name{blindingdata}
\docType{data}
\alias{bd010}
\alias{bd011}
\alias{bd012}

\title{Simulated blinding data}
\description{
 Simulated blinding survey datasets .
}
\format{Key variables
  \tabular{lll}{
    \code{y}  \tab vector  \tab outcome variable \cr
    \code{guess}  \tab vector  \tab guess of treatments \cr
    \code{group}  \tab vector  \tab group assignment
  }
}
\examples{
data(bd012)
names(bd012)
}
\keyword{datasets}
