\name{ex100x3}
\docType{data}
\alias{ex100x3}
\alias{n100x3}

\title{Sample PRO Data With Repeated Measures}
\description{
 A simulated data set of patient-reported outcomes with repeated measures.
}
\format{A data frame with observations at beaseline and at a follow-up time.
  \tabular{lll}{
    \code{w0}  \tab matrix  \tab measures at baseline \cr
    \code{w1}  \tab matrix  \tab measures at follow-up time \cr
    \code{group}  \tab character  \tab group assignment
  }
}
\keyword{datasets}
