% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prome-package.R
\docType{package}
\name{prome-package}
\alias{prome-package}
\alias{prome}
\title{The 'prome' package.}
\description{
  Algorithms to implenment the Bayesian methods to denoise the
  measurement errors in patient-reported outcome data with repeated
  measures. Also, two algorithms are included to discount the subgroup
  means or proportions for clinical studies with multiple subgroups.
}
